[test] move graph tests to e2e package
