+++ packages/tools/e2e-tests/.eslintrc.js
@@ -5,9 +5,17 @@ module.exports = {
     '@kadena-dev/eslint-config/profile/lib',
     'plugin:playwright/recommended',
   ],
+  "import/resolver": {
+    "typescript": {}
+  },
+  parser: "@typescript-eslint/parser",
   parserOptions: { tsconfigRootDir: __dirname },
   rules: {
     '@kadena-dev/typedef-var': 'off',
     '@rushstack/typedef-var': 'off',
   },
+  plugins: [
+    "@typescript-eslint",
+    "import"
+  ]
 };
