+++ packages/tools/e2e-tests/package.json
@@ -21,6 +21,10 @@
     "@kadena-dev/lint-package": "workspace:*",
     "@kadena-dev/markdown": "workspace:*",
     "@kadena-dev/shared-config": "workspace:*",
+    "@kadena/client": "workspace:*",
+    "@kadena/client-utils": "workspace:*",
+    "@kadena/cryptography-utils": "workspace:*",
+    "@kadena/types": "workspace:*",
     "@playwright/test": "^1.40.1",
     "@rushstack/eslint-config": "~3.3.0",
     "@types/node": "^18.17.14",
@@ -29,5 +33,8 @@
     "eslint-plugin-playwright": "0.16.0",
     "parse-md": "2.0.5",
     "prettier": "~3.0.3"
+  },
+  "dependencies": {
+    "eslint-import-resolver-typescript": "3.5.5"
   }
 }
