+++ packages/tools/e2e-tests/src/fixtures/graph/testdata/constants/amounts.ts
