+++ packages/tools/e2e-tests/src/fixtures/graph/testdata/queries/getAccount.ts
