+++ packages/tools/e2e-tests/src/helpers/docs/cookie.helper.ts
