+++ packages/tools/e2e-tests/src/helpers/docs/menu.helper.ts
