+++ packages/tools/e2e-tests/src/helpers/docs/mock.helper.ts
