+++ packages/tools/e2e-tests/src/helpers/graph/cryptography.helper.ts
