+++ packages/tools/e2e-tests/src/page-objects/index.ts
@@ -1,5 +1,5 @@
 import { test as baseTest } from '@playwright/test';
-import { MockHelper } from '../helpers/mock.helper';
+import { MockHelper } from '../helpers/docs/mock.helper';
 import { ToolsAppIndex } from './tools-app/toolsApp.index';
 
 export const test = baseTest.extend<{
