+++ packages/tools/e2e-tests/src/tests/docs-app/page-crawler.spec.ts
@@ -2,8 +2,8 @@
 /* eslint-disable playwright/expect-expect */
 // The expect is performed by validateTree(), so the test does not need one.
 import { test } from '@playwright/test';
-import CookieHelper from '../../helpers/cookie.helper';
-import TreeHelper from '../../helpers/tree.helper';
+import CookieHelper from '../../helpers/docs/cookie.helper';
+import TreeHelper from '../../helpers/docs/tree.helper';
 
 const pages = [
   'kadena',
